requirements.txt
fastapi
uvicorn
msal
azure-storage-blob
msgraph-core
msgraph-sdk